+++ src/algorithms/dreamer/rssm.py
@@ -8,7 +8,7 @@ import torch
 from torch import distributions as torchd
 from torch import nn
 
-import src.algorithms.dreamer.distributions as dists  #! R2Dreamer used bare `import distributions as dists`
+import src.components.distributions as dists  #! R2Dreamer used bare `import distributions as dists`
 from src.algorithms.dreamer.networks import BlockLinear, LambdaLayer  #!
 from src.algorithms.dreamer.tools import rpad, weight_init_  #!
 
@@ -22,17 +22,17 @@ class Deter(nn.Module):
         act = getattr(torch.nn, act)
         self._dyn_in0 = nn.Sequential(
             nn.Linear(deter, hidden, bias=True),
-            nn.RMSNorm(hidden, eps=1e-04, dtype=torch.float32),
+            nn.RMSNorm(hidden, eps=1e-04, dtype=torch.bfloat16),
             act(),
         )
         self._dyn_in1 = nn.Sequential(
             nn.Linear(stoch, hidden, bias=True),
-            nn.RMSNorm(hidden, eps=1e-04, dtype=torch.float32),
+            nn.RMSNorm(hidden, eps=1e-04, dtype=torch.bfloat16),
             act(),
         )
         self._dyn_in2 = nn.Sequential(
             nn.Linear(act_dim, hidden, bias=True),
-            nn.RMSNorm(hidden, eps=1e-04, dtype=torch.float32),
+            nn.RMSNorm(hidden, eps=1e-04, dtype=torch.bfloat16),
             act(),
         )
         self._dyn_hid = nn.Sequential()
@@ -42,7 +42,7 @@ class Deter(nn.Module):
                 f"dyn_hid_{i}", BlockLinear(in_ch, deter, self.blocks)
             )
             self._dyn_hid.add_module(
-                f"norm_{i}", nn.RMSNorm(deter, eps=1e-04, dtype=torch.float32)
+                f"norm_{i}", nn.RMSNorm(deter, eps=1e-04, dtype=torch.bfloat16)
             )
             self._dyn_hid.add_module(f"act_{i}", act())
             in_ch = deter
@@ -128,7 +128,7 @@ class RSSM(nn.Module):
             )
             self._obs_net.add_module(
                 f"obs_net_n_{i}",
-                nn.RMSNorm(self._hidden, eps=1e-04, dtype=torch.float32),
+                nn.RMSNorm(self._hidden, eps=1e-04, dtype=torch.bfloat16),
             )
             self._obs_net.add_module(f"obs_net_a_{i}", act())
             inp_dim = self._hidden
@@ -150,7 +150,7 @@ class RSSM(nn.Module):
             )
             self._img_net.add_module(
                 f"img_net_n_{i}",
-                nn.RMSNorm(self._hidden, eps=1e-04, dtype=torch.float32),
+                nn.RMSNorm(self._hidden, eps=1e-04, dtype=torch.bfloat16),
             )
             self._img_net.add_module(f"img_net_a_{i}", act())
             inp_dim = self._hidden
